add more information for our remotes
It was frustrating when I would attempt to rebuild the github repos and
I was getting information that they could not be deleted because I had
the wrong json output. I've added some automation to help me with that
along with make targets that touch the files I need to touch.