fix html processing
There was a fun little reminder that pandoc attempts to parse markdown
code in HTML, which is a setting we explicitly requested to allow for
lessons to use block-level HTML to express things they can not express
in markdown.

The problem is that when we have automated HTML output that places >=4
spaces between tags (e.g. in tables), then those spaces are interpreted
as markdown code blocks:

pandoc -t html -f markdown <<< '<tr>okay</tr>    <tr>not okay</tr>'

This was in some of the lessons, so we are processing these
possibilities here since we know that space does not matter in HTML.

see https://github.com/jgm/pandoc/issues/7209