make sure we force the main branch
In cases where there was a main branch and it was renamed to
`legacy/main`, we will need to force-push even after fetching and
pruning because there is a good chance that github will complain that
the branch had previously existed.

But, because we had preserved the main branch under `legacy/main`, we
feel confident that we can force it.