set install AND remove packages to eval=FALSE
It was driving me nuts trying to figure out why bioc-project would not
load the S4Vectors package for that particular episode until I found out
that it was _removing_ the S4Vectors package at the end of the preceding
episode.

I have set all of the code chunks that muck with the library state to
not evaluate.