[filter] keep setup markdown files from includes
Some overview lessons have setup files inside of the _includes folder.
The Workbench was not designed for this and we explicitly remove the
_includes folder on filtering.

This commit adds a catch during git-filter-repo that will rename any
setup markdown file in _includes/ to act like it was in the files/
folder all along.